Fix for bug 33421: Support null values for task priority.

+++ openexchange-test/unittests/com/openexchange/data/conversion/ical/ICalParserBasicTests.java
@@ -846,7 +846,7 @@ public class ICalParserBasicTests extends AbstractICalParserTest {
     private void priorityTest(final int priority, final int expected) throws ConversionError {
         final String icalText = fixtures.vtodoWithSimpleProperties("PRIORITY", new Integer(priority).toString());
         final Task task = parseTask(icalText);
-        assertEquals("Invalid interpretation for priority "+priority, expected, task.getPriority());
+        assertEquals("Invalid interpretation for priority " + priority, expected, task.getPriority().intValue());
     }
 
     // Bug 10401
